asyncKafka/requirements.txt
fastapi
uvicorn
kafka-python
transformers
streamlit
